requirements.txt
Flask==2.1.1
selenium==4.1.0
werkzeug==2.1.1
webdriver-manager-4.0.2